WOR-1491: make workflows_bindings_rpc obsolete
WOR-1537 added a compatibility-gate bypass for private inner
fetchers when constructing wrapped bindings. This allows Workflow
bindings on old compatibility dates to resolve their JSRPC methods.

Use JSRPC unconditionally and remove the legacy HTTP transport. Keep
the workflows_bindings_rpc flag name for configuration validation,
but make it a no-op. Cover old dates in workerd and edgeworker tests.
