+++ src/cloudflare/internal/test/workflows/BUILD.bazel
@@ -2,6 +2,5 @@ load("//:build/wd_test.bzl", "wd_test")
 
 wd_test(
     src = "workflows-api-test.wd-test",
-    args = ["--experimental"],
     data = glob(["*.js"]),
 )
