+++ src/cloudflare/internal/workflows-api.ts
@@ -11,16 +11,12 @@ export class NonRetryableError extends Error {
   }
 }
 
-const workflowsBindingsRpc =
-  !!Cloudflare.compatibilityFlags['workflows_bindings_rpc'];
-
 type WorkflowBatchDeleteResult = {
   deleted: { id: string }[];
   errors: { id: string; code: number; message: string }[];
 };
 
 interface Fetcher {
-  fetch: typeof fetch;
   getInstance(id: string): Promise<{ id: string }>;
   deleteInstance(id: string): Promise<void>;
   create(options?: WorkflowInstanceCreateOptions): Promise<{ id: string }>;
@@ -45,32 +41,6 @@ interface Fetcher {
   ): Promise<void>;
 }
 
-async function callFetcher<T>(
-  fetcher: Fetcher,
-  path: string,
-  body: object
-): Promise<T> {
-  const res = await fetcher.fetch(`http://workflow-binding.local${path}`, {
-    method: 'POST',
-    headers: {
-      'Content-Type': 'application/json',
-      'X-Version': '1',
-    },
-    body: JSON.stringify(body),
-  });
-
-  const response = (await res.json()) as {
-    result: T;
-    error?: WorkflowError;
-  };
-
-  if (res.ok) {
-    return response.result;
-  } else {
-    throw new Error(response.error?.message);
-  }
-}
-
 class InstanceImpl implements WorkflowInstance {
   readonly #fetcher: Fetcher;
   readonly id: string;
@@ -81,66 +51,29 @@ class InstanceImpl implements WorkflowInstance {
   }
 
   async pause(): Promise<void> {
-    if (workflowsBindingsRpc) {
-      await this.#fetcher.pause(this.id);
-      return;
-    }
-    await callFetcher(this.#fetcher, '/pause', {
-      id: this.id,
-    });
+    await this.#fetcher.pause(this.id);
   }
+
   async resume(): Promise<void> {
-    if (workflowsBindingsRpc) {
-      await this.#fetcher.resume(this.id);
-      return;
-    }
-    await callFetcher(this.#fetcher, '/resume', {
-      id: this.id,
-    });
+    await this.#fetcher.resume(this.id);
   }
 
   async terminate(options?: WorkflowInstanceTerminateOptions): Promise<void> {
-    if (workflowsBindingsRpc) {
-      await this.#fetcher.terminate(this.id, options);
-      return;
-    }
-    await callFetcher(this.#fetcher, '/terminate', {
-      id: this.id,
-      ...(options?.rollback === true ? { rollback: true } : {}),
-    });
+    await this.#fetcher.terminate(this.id, options);
   }
 
   async restart(options?: WorkflowInstanceRestartOptions): Promise<void> {
-    if (workflowsBindingsRpc) {
-      await this.#fetcher.restart(this.id, options);
-      return;
-    }
-    await callFetcher(this.#fetcher, '/restart', {
-      ...options,
-      id: this.id,
-    });
+    await this.#fetcher.restart(this.id, options);
   }
 
   async delete(): Promise<void> {
-    if (workflowsBindingsRpc) {
-      // deleteInstance, not delete: avoids colliding with the built-in Fetcher.delete(url), which
-      // is still present on compatibility dates before `fetcher_no_get_put_delete`.
-      await this.#fetcher.deleteInstance(this.id);
-      return;
-    }
-    await callFetcher(this.#fetcher, '/delete', {
-      id: this.id,
-    });
+    // deleteInstance, not delete: avoids colliding with the built-in Fetcher.delete(url), which
+    // is still present on compatibility dates before `fetcher_no_get_put_delete`.
+    await this.#fetcher.deleteInstance(this.id);
   }
 
   async status(): Promise<InstanceStatus> {
-    if (workflowsBindingsRpc) {
-      return await this.#fetcher.status(this.id);
-    }
-    const result = await callFetcher<InstanceStatus>(this.#fetcher, '/status', {
-      id: this.id,
-    });
-    return result;
+    return await this.#fetcher.status(this.id);
   }
 
   async sendEvent({
@@ -150,15 +83,7 @@ class InstanceImpl implements WorkflowInstance {
     type: string;
     payload: unknown;
   }): Promise<void> {
-    if (workflowsBindingsRpc) {
-      await this.#fetcher.sendEvent(this.id, { type, payload });
-      return;
-    }
-    await callFetcher(this.#fetcher, '/send-event', {
-      type,
-      payload,
-      id: this.id,
-    });
+    await this.#fetcher.sendEvent(this.id, { type, payload });
   }
 }
 
@@ -171,12 +96,9 @@ class WorkflowImpl extends wrappedBinding.WrappedBinding {
   }
 
   async get(id: string): Promise<WorkflowInstance> {
-    const result = workflowsBindingsRpc
-      ? // getInstance, not get: avoids colliding with the built-in Fetcher.get(url).
-        await this.#fetcher.getInstance(id)
-      : await callFetcher<{
-          id: string;
-        }>(this.#fetcher, '/get', { id });
+    // getInstance, not get: avoids colliding with the built-in Fetcher.get(url), which is still
+    // present on compatibility dates before `fetcher_no_get_put_delete`.
+    const result = await this.#fetcher.getInstance(id);
 
     return new InstanceImpl(result.id, this.#fetcher);
   }
@@ -184,11 +106,7 @@ class WorkflowImpl extends wrappedBinding.WrappedBinding {
   async create(
     options?: WorkflowInstanceCreateOptions
   ): Promise<WorkflowInstance> {
-    const result = workflowsBindingsRpc
-      ? await this.#fetcher.create(options)
-      : await callFetcher<{
-          id: string;
-        }>(this.#fetcher, '/create', options ?? {});
+    const result = await this.#fetcher.create(options);
 
     return new InstanceImpl(result.id, this.#fetcher);
   }
@@ -196,26 +114,13 @@ class WorkflowImpl extends wrappedBinding.WrappedBinding {
   async createBatch(
     options: WorkflowInstanceCreateOptions[]
   ): Promise<WorkflowInstance[]> {
-    const results = workflowsBindingsRpc
-      ? await this.#fetcher.createBatch(options)
-      : await callFetcher<
-          {
-            id: string;
-          }[]
-        >(this.#fetcher, '/createBatch', options);
+    const results = await this.#fetcher.createBatch(options);
 
     return results.map((result) => new InstanceImpl(result.id, this.#fetcher));
   }
 
   async deleteBatch(instanceIds: string[]): Promise<WorkflowBatchDeleteResult> {
-    const options = { instances: instanceIds };
-    return workflowsBindingsRpc
-      ? await this.#fetcher.deleteBatch(options)
-      : await callFetcher<WorkflowBatchDeleteResult>(
-          this.#fetcher,
-          '/deleteBatch',
-          options
-        );
+    return await this.#fetcher.deleteBatch({ instances: instanceIds });
   }
 }
